Move the LicenseStatus enum
